% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{vhs}
\alias{vhs}
\title{Choose a VHS palette}
\usage{
vhs(name, n, type = "d")
}
\arguments{
\item{name}{name of VHS palette}

\item{n}{Number of colours required from palette}

\item{type}{Discrete "d" or continuous "c" (default - "c")}
}
\value{
A vector of colours
}
\description{
Choose a VHS palette
}
\examples{
vhs("vhs")
}
